% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgb.R
\name{cimg2rgb}
\alias{cimg2rgb}
\title{function converts 4d image data to dataframe}
\usage{
cimg2rgb(cimg)
}
\arguments{
\item{cimg}{An 4d image tensor}
}
\value{
A 2d dataframe with columns rgb.
}
\description{
function converts 4d image data to dataframe
}
\examples{
stadium = Photopal::stadium
cimg2rgb(stadium)
}
\author{
Hiiiua
}
