% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgb.R
\name{cimgdf2df}
\alias{cimgdf2df}
\title{function converts 4d image data to dataframe}
\usage{
cimgdf2df(cimg)
}
\arguments{
\item{cimg}{An 4d image tensor}
}
\value{
A 2d dataframe with columns rgb.
}
\description{
function converts 4d image data to dataframe
}
