% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgb.R
\name{color_blindness_palette}
\alias{color_blindness_palette}
\title{create color palette from color-blind simullation images}
\usage{
color_blindness_palette(
  loc = "stadium.rda",
  mode = "red",
  num.color = 5,
  df.rgb,
  threshold = 25,
  plot = T
)
}
\arguments{
\item{loc}{A string.}

\item{mode}{A string.}

\item{num.color}{An integer.}

\item{df.rgb}{A dataframe.}

\item{threshold}{A number.}

\item{plot}{T/F.}
}
\description{
create color palette from color-blind simullation images
}
