% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgb.R
\name{color_blindness_simulation}
\alias{color_blindness_simulation}
\title{different color-blind simulations}
\usage{
color_blindness_simulation(loc = "stadium.rda", mode = "red", compare = T)
}
\arguments{
\item{loc}{A string. A location of an image}

\item{mode}{A string. Specify color blind type}

\item{compare}{T/F. True for displaying both normal and color-blind vision to compare.}
}
\description{
This function takes in an image, transforming it to simulate color-blind vision
}
\examples{
color_blindness_simulation()
}
