% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgb.R
\name{contrast}
\alias{contrast}
\title{Determine the difference of two colors}
\usage{
contrast(c1, c2, maxColorValue = 255)
}
\arguments{
\item{c1}{A numeric list, rgb values of color1.}

\item{c2}{A numeric list, rgb values of color2.}

\item{maxColorValue}{Maximum color value, mostly 1 or 255}
}
\description{
This function returns the colors difference.
Passing in two sets of rgb values of two colors, gets returned
delta e color distance and display.
}
\examples{
contrast(c(70, 50, 50), c(0,0,0))
contrast(c(0.2745098, 0.1960784, 0.1960784),c(0,0,0), maxColorValue = 1)
}
