% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{image}
\alias{image}
\title{A Nebraska football stadium image data}
\format{
\subsection{\code{image}}{

A data of 4 dimensions, with 1500 rows, 1000 columns, 1 depth and 3 color channels.
}
}
\source{
\url{https://topbrunchspots.com/wp-content/uploads/2022/05/AF1QipM-ql78yz46DGynt2tb1l7i8gK8zOAfN8dCQqiPw1600-h1000-k-no.jpeg}
}
\usage{
image
}
\description{
An 4 dimensional image data
}
\keyword{datasets}
