% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgb.R
\name{image2rgb}
\alias{image2rgb}
\title{A function read in an image from url, locally or default build-in image.}
\usage{
image2rgb(
  url = paste("https://topbrunchspots.com", "/wp-content/uploads/2022/05/AF1QipM",
    "-ql78yz46DGynt2tb1l7i8gK8zOAfN8dCQqiP", "w1600-h1000-k-no.jpeg", sep = "")
)
}
\arguments{
\item{url}{A string.}
}
\description{
A function read in an image from url, locally or default build-in image.
}
