% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgb.R
\name{imgdf2df}
\alias{imgdf2df}
\title{function converts 4d image data to dataframe}
\usage{
imgdf2df(cimg)
}
\arguments{
\item{cimg}{An 4d image tensor}
}
\description{
function converts 4d image data to dataframe
}
