% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgb.R
\name{is_sufficient}
\alias{is_sufficient}
\title{Tell if the difference is sufficient for human eye to perceive}
\usage{
is_sufficient(c1, c2, maxColorValue = 255, threshold = 35)
}
\arguments{
\item{c1}{A numeric list, rgb values of color1.}

\item{c2}{A numeric list, rgb values of color2.}

\item{maxColorValue}{Maximum color value, mostly 1 or 255}

\item{threshold}{A number, threshold for differentiating two colors in delta e.}
}
\description{
This function returns whether the colors difference is sufficient for human eye
to detect or not.
}
\examples{
is_sufficient(c(70, 50, 50), c(0,0,0))
}
