% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stadium}
\alias{stadium}
\title{A Nebraska football stadium image data}
\format{
A data of 4 dimensions, with 1806 rows, 605 columns, 1 depth and 3 color channels.
}
\source{
\url{https://github.com/TWiedRW}
}
\usage{
stadium
}
\description{
An 4 dimensional image data
}
\keyword{datasets}
