% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgb.R
\name{test_function}
\alias{test_function}
\title{A testing function}
\usage{
test_function(x)
}
\arguments{
\item{x}{A number.}
}
\description{
\code{test_function()} prints input number.
This is just a testing function to make sure the package structure works.
}
\examples{
test_function(1)
}
