% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgb.R
\name{image2rgb}
\alias{image2rgb}
\title{A function read in an image from url, locally or default build-in image.}
\usage{
image2rgb(loc = "stadium.rda")
}
\arguments{
\item{loc}{A string. A valid url or file for the image}
}
\value{
A 2d dataframe with columns rgb.
}
\description{
A function read in an image from url, locally or default build-in image.
}
\examples{
im = image2rgb()
head(im)
}
